# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Game-Baza API
  description: 'REST API платформы для сравнения цен на игры. Поддерживает каталог игр, фильтрацию, аутентификацию через JWT-cookie.'
  contact:
    name: Game-Baza Team
    url: 'https://github.com/bembelem/game-baza'
    email: support@game-baza.example.com
  license:
    name: MIT
    url: 'https://opensource.org/licenses/MIT'
  version: 0.1.0
servers:
  - url: 'http://localhost:8080'
    description: Локальная разработка
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Регистрация
      description: Создаёт нового пользователя и устанавливает JWT-cookie с access_token.
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegisterRequest'
            examples:
              '1':
                summary: Обычный пользователь
                value:
                  username: john doe
                  email: john@example.com
                  password: SecurePass123!
                  birthdate: '2000-03-30'
      responses:
        '201':
          description: Пользователь успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPrivate'
        '409':
          description: Пользователь уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
              examples:
                email_exists:
                  summary: Email уже занят
                  value:
                    error: 2020_VALIDATION_ERROR
                    message: Данные не прошли валидацию.
                    details:
                      email: Пользователь с таким email уже существует.
                    traceId: 9696004a-6866-4a79-bc53-f4383040cba1
                username_exists:
                  summary: Username уже занят
                  value:
                    error: 2020_VALIDATION_ERROR
                    message: Данные не прошли валидацию.
                    details:
                      username: Пользователь с таким username уже существует.
                    traceId: 9696004a-6866-4a79-bc53-f4383040cba1
  /auth/login:
    post:
      tags: [Auth]
      summary: Вход
      description: 'Аутентифицирует пользователя по email или username, устанавливает JWT-cookie.'
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
            examples:
              '1':
                summary: Вход по email
                value:
                  email: john@example.com
                  password: SecurePass123!
              '2':
                summary: Вход по username
                value:
                  username: john doe
                  password: SecurePass123!
      responses:
        '200':
          description: Успешный вход
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPrivate'
        '401':
          description: Ошибка авторизации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
        '422':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Выход
      description: Удаляет JWT-cookie и завершает сессию пользователя.
      operationId: logoutUser
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Успешный выход
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
...
  /games:
    get:
      tags: [Games]
      summary: Список игр
      description: Возвращает постраничный список игр с минимальными ценами. Поддерживает фильтрацию и курсорную пагинацию.
      operationId: listGames
      parameters:
        - name: last_id
          in: query
          required: false
          description: ID последней записи (cursor)
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: per_page
          in: query
          required: false
          description: Записей на страницу
          schema:
            type: integer
            maximum: 20
            exclusiveMinimum: 0
            default: 20
        - name: title
          in: query
          required: false
          description: Поиск по названию игры
          schema:
            type: string
            maxLength: 100
        - name: stores
          in: query
          required: false
          description: Фильтр по магазинам
          schema:
            type: array
            items:
              type: string
        - name: platforms
          in: query
          required: false
          description: Фильтр по платформам
          schema:
            type: array
            items:
              type: string
        - name: genres
          in: query
          required: false
          description: Фильтр по жанрам
          schema:
            type: array
            items:
              type: string
        - name: price_min
          in: query
          required: false
          description: Минимальная цена в рублях
          schema:
            type: integer
            minimum: 0
            maximum: 999999
        - name: price_max
          in: query
          required: false
          description: Максимальная цена в рублях
          schema:
            type: integer
            minimum: 0
            maximum: 999999
        - name: sort
          in: query
          required: false
          description: 'Сортировка: price_asc, price_desc, title_asc, title_desc'
          schema:
            type: string
            enum: [price_asc, price_desc, title_asc, title_desc]
      responses:
        '200':
          description: Постраничный список игр
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GamesResponse'
        '422':
          description: Ошибка валидации параметров
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
  '/games/{game_id}':
    get:
      tags: [Games]
      summary: Информация об игре
      description: 'Возвращает детальную информацию об игре: описание, жанры, разработчика, издателя и офферы по магазинам.'
      operationId: getGame
      parameters:
        - name: game_id
          in: path
          required: true
          description: ID игры
          schema:
            type: integer
      responses:
        '200':
          description: Детальная информация об игре
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameDetails'
        '404':
          description: Игра не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
...
  '/stores/{store_id}':
    get:
      tags: [Stores]
      summary: Магазин
      description: Возвращает магазин по его ID.
      operationId: getStore
      parameters:
        - name: store_id
          in: path
          required: true
          description: ID магазина
          schema:
            type: integer
      responses:
        '200':
          description: Данные магазина
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Store'
        '404':
          description: Магазин не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
...
  '/platforms/{platform_id}':
    get:
      tags: [Platforms]
      summary: Платформа
      description: Возвращает платформу по её ID.
      operationId: getPlatform
      parameters:
        - name: platform_id
          in: path
          required: true
          description: ID платформы
          schema:
            type: integer
      responses:
        '200':
          description: Данные платформы
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Platform'
        '404':
          description: Платформа не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
...
  '/publishers/{publisher_id}':
    get:
      tags: [Publishers]
      summary: Издатель
      description: Возвращает издателя по его ID.
      operationId: getPublisher
      parameters:
        - name: publisher_id
          in: path
          required: true
          description: ID издателя
          schema:
            type: integer
      responses:
        '200':
          description: Данные издателя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Publisher'
        '404':
          description: Издатель не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
...
  '/developers/{developer_id}':
    get:
      tags: [Developers]
      summary: Разработчик
      description: Возвращает разработчика по его ID.
      operationId: getDeveloper
      parameters:
        - name: developer_id
          in: path
          required: true
          description: ID разработчика
          schema:
            type: integer
      responses:
        '200':
          description: Данные разработчика
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Developer'
        '404':
          description: Разработчик не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
...
  '/genres/{genre_id}':
    get:
      tags: [Genres]
      summary: Жанр
      description: Возвращает жанр по его ID.
      operationId: getGenre
      parameters:
        - name: genre_id
          in: path
          required: true
          description: ID жанра
          schema:
            type: integer
      responses:
        '200':
          description: Данные жанра
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Genre'
        '404':
          description: Жанр не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: JWT-токен в HttpOnly cookie. Устанавливается /auth/login и /auth/register.
  schemas:
...
    UserRegisterRequest:
      type: object
      title: UserRegisterRequest
      description: Тело запроса POST /auth/register.
      required: [username, email, birthdate, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          pattern: '^[a-zA-Z0-9_ ]+$'
          description: '3-50 символов: латиница, цифры, подчёркивания, пробелы'
        email:
          type: string
          format: email
          description: Email пользователя
        birthdate:
          type: string
          format: date
          description: Дата рождения (YYYY-MM-DD)
        password:
          type: string
          minLength: 6
          maxLength: 50
          pattern: '^(?=.*\d)(?=.*[!@#$%^&*()_\-+=\[\]{};:''\",.<>/?\\|`~]).+$'
          description: 'Минимум 6 символов, хотя бы одна цифра и один спецсимвол'
      example:
        username: john_doe
        email: john@example.com
        birthdate: '2000-01-15'
        password: SecurePass123!
    UserLoginRequest:
      type: object
      title: UserLoginRequest
      description: Тело запроса POST /auth/login. Нужно указать либо email, либо username.
      required: [password]
      properties:
        username:
          type: string
          nullable: true
        email:
          type: string
          format: email
          nullable: true
        password:
          type: string
      example:
        email: john@example.com
        password: SecurePass123!

    # ----- Users -----
...
    UserPrivate:
      type: object
      title: UserPrivate
      description: Полный профиль пользователя — отдаётся только владельцу (/users/me).
      required: [id, username, email, birthdate, created_at]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
          format: email
        birthdate:
          type: string
          format: date
        created_at:
          type: string
          format: date
      example:
        id: 42
        username: john_doe
        email: john@example.com
        birthdate: '2000-01-15'
        created_at: '2024-05-20'
...
    GameCard:
      type: object
      title: GameCard
      description: Карточка игры для списка (response для GET /games).
      required: [id, title, image_url, min_price_original, min_price_discount, discount_percent]
      properties:
        id:
          type: integer
        title:
          type: string
        image_url:
          type: string
          format: uri
        min_price_original:
          type: integer
          description: Минимальная цена без скидки среди всех офферов
        min_price_discount:
          type: integer
          description: Минимальная цена со скидкой среди всех офферов
        discount_percent:
          type: integer
          minimum: 0
          maximum: 100
      example:
        id: 1
        title: Cyberpunk 2077
        image_url: 'https://example.com/cyberpunk.jpg'
        min_price_original: 2999
        min_price_discount: 1499
        discount_percent: 50
    GameDetails:
      type: object
      title: GameDetails
      description: 'Подробная страница игры (response для GET /games/{id}).'
      required: [id, title, image_url, min_price_original, min_price_discount, discount_percent]
      properties:
        id:
          type: integer
        title:
          type: string
        image_url:
          type: string
          format: uri
        min_price_original:
          type: integer
        min_price_discount:
          type: integer
        discount_percent:
          type: integer
          minimum: 0
          maximum: 100
        description:
          type: string
          nullable: true
        release_date:
          type: string
          format: date
          nullable: true
        developer:
          type: string
          nullable: true
        publisher:
          type: string
          nullable: true
        genres:
          type: array
          items:
            type: string
          default: []
        offers:
          type: array
          items:
            $ref: '#/components/schemas/Offer'
          default: []
      example:
        id: 1
        title: Cyberpunk 2077
        image_url: 'https://example.com/cyberpunk.jpg'
        min_price_original: 2999
        min_price_discount: 1499
        discount_percent: 50
        description: 'Cyberpunk 2077 — это приключенческая RPG в Найт-Сити.'
        release_date: '2020-12-10'
        developer: CD Projekt Red
        publisher: CD Projekt
        genres: [Экшены, РПГ]
        offers:
          - id: 10
            game_id: 1
            store_id: 1
            store: Steam
            store_game_link: 'https://store.steampowered.com/app/1091500/'
            price_original: 2999
            price_discount: 1499
            discount_percent: 50
            positive_percent: 85
    GamesResponse:
      type: object
      title: GamesResponse
      description: Постраничный ответ каталога — список карточек + метаданные.
      required: [total, last_id, per_page, has_more, items]
      properties:
        total:
          type: integer
          minimum: 0
        last_id:
          type: integer
          nullable: true
          description: ID последнего элемента (cursor для следующей страницы)
        per_page:
          type: integer
        has_more:
          type: boolean
        items:
          type: array
          items:
            $ref: '#/components/schemas/GameCard'
      example:
        total: 5322
        last_id: 20
        per_page: 20
        has_more: true
        items:
          - id: 1
            title: Cyberpunk 2077
            image_url: 'https://example.com/cyberpunk.jpg'
            min_price_original: 2999
            min_price_discount: 1499
            discount_percent: 50

    # ----- Offers -----
    Offer:
      type: object
      title: Offer
      required: [id, game_id, store_id, store, store_game_link, price_original, price_discount, discount_percent, positive_percent]
      properties:
        id:
          type: integer
        game_id:
          type: integer
        store_id:
          type: integer
        store:
          type: string
          description: Название магазина
        store_game_link:
          type: string
          format: uri
        price_original:
          type: integer
        price_discount:
          type: integer
        discount_percent:
          type: integer
          minimum: 0
          maximum: 100
        positive_percent:
          type: integer
          minimum: 0
          maximum: 100
          description: Процент положительных отзывов в магазине
      example:
        id: 10
        game_id: 1
        store_id: 1
        store: Steam
        store_game_link: 'https://store.steampowered.com/app/1091500/'
        price_original: 2999
        price_discount: 1499
        discount_percent: 50
        positive_percent: 85
...
    Genre:
      type: object
      title: Genre
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
      example:
        id: 1
        name: Экшены
...
    Store:
      type: object
      title: Store
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        url:
          type: string
          format: uri
          nullable: true
      example:
        id: 1
        name: Steam
        url: 'https://store.steampowered.com/'
...
    Platform:
      type: object
      title: Platform
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        url:
          type: string
          format: uri
          nullable: true
      example:
        id: 1
        name: PC
        url: null
...
    Publisher:
      type: object
      title: Publisher
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
      example:
        id: 1
        name: CD Projekt
...
    Developer:
      type: object
      title: Developer
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
      example:
        id: 1
        name: CD Projekt Red
...
    MessageResponse:
      type: object
      title: MessageResponse
      required: [message]
      properties:
        message:
          type: string
      example:
        message: Профиль удалён
    ErrorDetailResponse:
      type: object
      title: ErrorDetailResponse
      description: Унифицированный формат всех ошибок API.
      required: [error, message, details, traceId]
      properties:
        error:
          type: string
          description: Машино-читаемый код ошибки (например 2010_AUTHENTICATION_ERROR)
        message:
          type: string
          description: Человеко-читаемое сообщение
        details:
          type: object
          additionalProperties: true
          description: Детали ошибки по полям
        traceId:
          type: string
          description: ID трассировки запроса (X-Request-ID), для дебага
      example:
        error: 2010_AUTHENTICATION_ERROR
        message: Ошибка авторизации.
        details:
          password: Неверный пароль.
        traceId: 9696004a-6866-4a79-bc53-f4383040cba1

  # ----- Error Responses -----
...